+++ drools-framework-runtime-base/src/test/java/org/chtijbug/drools/runtime/test/RuleHistoryEventTest.java
@@ -38,7 +38,7 @@ public class RuleHistoryEventTest {
             }
         };
         RuleBasePackage ruleBasePackage = RuleBaseBuilder.newRuleBasePackage(historyListener, "com.pymmasoftware.test", "fibonacci", Arrays.asList("fibonacci.drl"));
-        int rulePackageID = ruleBasePackage.getRuleBaseID();
+        Long rulePackageID = ruleBasePackage.getRuleBaseID();
 
         RuleBaseSession ruleBaseSession1 = ruleBasePackage.createRuleBaseSession();
         Fibonacci fibonacci = new Fibonacci(1);
@@ -78,7 +78,7 @@ public class RuleHistoryEventTest {
             }
         };
         RuleBasePackage ruleBasePackage = RuleBaseBuilder.newRuleBasePackage(historyListener, "com.pymmasoftware.test", "fibonacci", Arrays.asList("ruleflow2.drl", "RuleFlowProcess2.bpmn2"));
-        int rulePackageID = ruleBasePackage.getRuleBaseID();
+        Long rulePackageID = ruleBasePackage.getRuleBaseID();
 
         RuleBaseSession ruleBaseSession1 = ruleBasePackage.createRuleBaseSession();
         Fibonacci fibonacci = new Fibonacci(0);
