+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/process/AfterProcessStartHistoryEvent.java
@@ -29,7 +29,7 @@ public class AfterProcessStartHistoryEvent extends ProcessHistoryEvent {
     public AfterProcessStartHistoryEvent() {
     }
 
-    public AfterProcessStartHistoryEvent(Long eventID, DroolsProcessInstanceObject processInstance, int ruleBaseId, int sessionId) {
+    public AfterProcessStartHistoryEvent(Long eventID, DroolsProcessInstanceObject processInstance, Long ruleBaseId, int sessionId) {
         super(eventID, ruleBaseId, sessionId);
         this.processInstance = processInstance;
     }
