+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/process/BeforeNodeInstanceTriggeredHistoryEvent.java
@@ -33,7 +33,7 @@ public class BeforeNodeInstanceTriggeredHistoryEvent extends ProcessHistoryEvent
     public BeforeNodeInstanceTriggeredHistoryEvent() {
     }
 
-    public BeforeNodeInstanceTriggeredHistoryEvent(Long eventID, DroolsNodeInstanceObject nodeInstance, int ruleBaseId, int sessionId) {
+    public BeforeNodeInstanceTriggeredHistoryEvent(Long eventID, DroolsNodeInstanceObject nodeInstance, Long ruleBaseId, int sessionId) {
         super(eventID, ruleBaseId, sessionId);
         this.nodeInstance = nodeInstance;
     }
