+++ drools-framework-runtime-entity/src/main/java/org/chtijbug/drools/entity/history/rule/AfterRuleFlowActivatedHistoryEvent.java
@@ -28,7 +28,7 @@ public class AfterRuleFlowActivatedHistoryEvent extends RuleFlowHistoryEvent {
     public AfterRuleFlowActivatedHistoryEvent() {
     }
 
-    public AfterRuleFlowActivatedHistoryEvent(Long eventID, DroolsRuleFlowGroupObject droolsRuleFlowGroupObject, int ruleBaseId, int sessionId) {
+    public AfterRuleFlowActivatedHistoryEvent(Long eventID, DroolsRuleFlowGroupObject droolsRuleFlowGroupObject, Long ruleBaseId, int sessionId) {
         super(eventID, droolsRuleFlowGroupObject, ruleBaseId, sessionId);
     }
